make legacy tests working with pytest
